# HEIGVD-AMT-Gamification/spring-server/src/main/resources/api-spec.yaml
swagger: '2.0'
info:
  title: Gamification API
  description: Move your app forward with the RFPS API.
  version: "1.0.0"
# the domain of the service
host: localhost:8080
...
schemes:
  - http
  - https
# will be prefixed to all paths
basePath: /v1
paths:
...
  /rules:
    get:
      description: 
        The rules of the application whose id is present in the authentication token.
      produces:
        - application/json
      parameters:
        - name: authenticationToken
          in: header
          description: Token used to authenticate the client application. Token must be preceded by "Bearer " pattern.
          required: true
          type: string
      tags:
        - Rules
      responses:
        200:
          description: OK, an array of rules.
          schema:
            type: array
            items:
              $ref: '#/definitions/RuleOutputDTO'
        400:
          description: Bad request, if Authorization header isn't provided or if it's not properly named.
        401:
          description: Unauthorized, if authentication token provided in the Authorization header is malformed, not signed by the gamification API server or expired. 
        500:
          description: Internal server error, if database isn't reachable.
              
    post:
      description: Create a new rule for the application whose id is present in the authentication token.
      consumes:
        - application/json
      parameters:
        - in: body
          name: rule
          description: Rule object that needs to be added to the database.
          required: true
          schema:
              $ref: '#/definitions/RuleInputDTO'
        - name: authenticationToken
          in: header
          description: Token used to authenticate the client application. Token must be preceded by "Bearer " pattern.
          required: true
          type: string
      responses:
        201:
          description: Created, returns the URL to access the new rule.
          schema:
            $ref: '#/definitions/LocationRule'
        400:
          description: Bad request, if Authorization header isn't provided, if it's not properly named or if json payload is malformed.
        401:
          description: Unauthorized, if authentication token provided in the Authorization header is malformed, not signed by the gamification API server or expired. 
        422:
          description: Unprocessable entity,
                       if mandatory fields aren't provided,
                       if mandatory fields provided are empty or contain only spaces,
                       if name provided contains more than 80 characters,
                       if description or imageUrl provided contain more than 255 characters,
                       if rule name provided already exists.
        500:
          description: Internal server error, if database isn't reachable.
  /rules/{id}:
    get:
      description: Returns the rule which has the specified id.
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          type: integer
          format: int64
          description: The id of the desired rule.
          required: true
        - name: authenticationToken
          in: header
          description: Token used to authenticate the client application. Token must be preceded by "Bearer " pattern.
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/RuleOutputDTO'
        400:
          description: Bad request, if Authorization header isn't provided or if it's not properly named.
        401:
          description: Unauthorized, if authentication token provided in the Authorization header is malformed, not signed by the gamification API server or expired. 
        404:
          description: Not found, if id provided doesn't exist.
        500:
          description: If database isn't reachable.
          
    put:
      description: Modify completely the existing rule which has the specified id.
      consumes:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the rule to modify.
          required: true
          type: integer
          format: int64
        - name: rule
          in: body
          description: rule to be updated
          required: true
          schema:
              $ref: '#/definitions/RuleInputDTO'
        - name: authenticationToken
          in: header
          description: Token used to authenticate the client application. Token must be preceded by "Bearer " pattern.
          required: true
          type: string
      responses:
        204:
          description: No content, if success.
        404:
          description: Not found, if id provided doesn't exist.
        400:
          description: Bad request, if Authorization header isn't provided, if it's not properly named or if json payload is malformed.
        401:
          description: Unauthorized, if authentication token provided in the Authorization header is malformed, not signed by the gamification API server or expired. 
        422:
          description: Unprocessable entity,
                       if mandatory fields aren't provided,
                       if mandatory fields provided are empty or contain only spaces,
                       if name provided contains more than 80 characters,
                       if description or imageUrl provided contain more than 255 characters,
                       if rule name provided already exists.
        500:
          description: Internal server error, if database isn't reachable.

    delete:
      description: Delete the rule which has the specified id.
      parameters:
        - name: id
          in: path
          description: The id of the desired rule.
          required: true
          type: integer
          format: int64
        - name: authenticationToken
          in: header
          description: Token used to authenticate the client application. Token must be preceded by "Bearer " pattern.
          required: true
          type: string
      responses:
        204:
          description: No content, if success.
        400:
          description: Bad request, if Authorization header isn't provided or if it's not properly named.
        401:
          description: Unauthorized, if authentication token provided in the Authorization header is malformed, not signed by the gamification API server or expired. 
        404:
          description: Not found, if id provided doesn't exist.
        500:
          description: Internal server error, if database isn't reachable.
definitions:
...
  RuleOutputDTO:
    type: object
    required: [rule_id, name, description, points]
    properties:
      rule_id:
        type: string
        description: Unique identifier representing a specific rule.
      name:
        type: string
        description: Display rule's name.
      description:
        type: string
        description: Description of rule.
      points:
        type: integer
        description: Number of points to get or lose.
  RuleInputDTO:
    type: object
    required: [ruleName, description, badge_id, pointScale_id, points, eventType]
    properties:
      ruleName:
        type: string
        description: Display rule's name.
      description:
        type: string
        description: Description of the rule.
      badge_id:
        type: integer
        description: The id of the badge to be given to a user
      pointScale_id:
        type: integer
        description: The id of the pointScale to be given to a user
      points:
        type: integer
        description: score of point to get/lose by this rule
      eventType:
        type: string
        description: The type of the up coming event
  LocationRule:
      type: string
      required: [locationRule]
      properties:
        locationRule:
          type: string
          description: Unique URL to access the new object created.
...